Add stats.
