+++ Shared/Samples/Animate 3D graphic/README.metadata.json
@@ -47,7 +47,8 @@
         "SurfacePlacement"
     ],
     "snippets": [
-        "Animate3DGraphicView.swift"
+        "Animate3DGraphicView.swift",
+        "Animate3DGraphicView.Model.swift"
     ],
     "title": "Animate 3D graphic"
 }
